backend/requirements.txt
cachetools==6.1.0
fastapi==0.115.14
fastapi_limiter==0.1.6
langchain_core==0.3.66
langchain_openai==0.3.27
pydantic==2.11.7
python-dotenv==1.1.1
redis==6.2.0
stytch==13.9.0
